Update schema_deployment_process.go
Fixes #127

+++ octopusdeploy/schema_deployment_process.go
@@ -21,7 +21,7 @@ func expandDeploymentProcess(d *schema.ResourceData) *octopusdeploy.DeploymentPr
 	}
 
 	if v, ok := d.GetOk("version"); ok {
-		deploymentProcess.Version = v.(int32)
+		deploymentProcess.Version = int32(v.(int))
 	}
 
 	if v, ok := d.GetOk("step"); ok {
